fix(dsv4/long-ctx): cast row indexing to int64 in long-context prefill kernels
Three kernels in the DSV4 prefill stack used int32 ``row * stride`` pointer
math that wraps under long-context (≥1M token) prefill. With CP4 zigzag a
1.1M sequence yields ~274092 rank-local Q rows; combined with strides
8565 / 9600 (combine_topk SWA) and 128 (per_token_group_quant block),
the product crosses 2^31 and writes off-allocation. The error surfaces as
a sticky CUDA_ERROR_ILLEGAL_ADDRESS at the next sync — typically inside
DeepGEMM's TMA encode or flash_mla_sparse_fwd, masking the real culprit.

- ``per_token_group_quant_8bit_v2.cu``: promote ``offset_num_groups`` to
  int64 so ``offset_num_groups * GROUP_SIZE`` cannot wrap. Reproduced at
  M=274092 K=8192 (DSV4 wo_b).
- ``_swa_ops_triton.py::_combine_topk_swa_indices_kernel`` and
  ``_combine_topk_swa_indices_cp_kernel``: cast the row index to
  ``tl.int64`` once before the pointer expressions (both load and store
  paths) so ``row * topk_indices_stride`` and ``row * combined_indices_stride``
  stay in int64.
- ``Attention._attn_via_workspace``: chunk Q across ``flash_mla_sparse_fwd``
  launches (default 65536, env ``DSV4_FLASH_MLA_SPARSE_Q_CHUNK``). Defensive
  — flash_mla v1.0.0 has not been audited for int32 indexing at the same
  scale; chunking is bit-equal because sparse attention has no cross-Q
  dependency.

Verified: LongBench-V2 q62 (~1.09M tokens) under PD-disagg CP4 prefill +
TP1DP1EP1 decode now returns the correct answer in ~3 minutes.

+++ rtp_llm/models_py/bindings/cuda/kernels/per_token_group_quant_8bit_v2.cu
@@ -249,8 +249,8 @@ __global__ void per_token_group_quant_8bit_kernel(const T* __restrict__ input,
                                                   DST_DTYPE* __restrict__ output_q,
                                                   scale_packed_t* __restrict__ output_s,
                                                   const int32_t* __restrict__ masked_m,
-                                                  const int subwarps_per_block,
-                                                  const int hidden_dim_num_groups,
+                                                  const int   subwarps_per_block,
+                                                  const int   hidden_dim_num_groups,
                                                   const float eps,
                                                   // TODO can this be removed?
                                                   const int scale_expert_stride,
@@ -273,8 +273,20 @@ __global__ void per_token_group_quant_8bit_kernel(const T* __restrict__ input,
             constexpr uint32_t INPUT_PRIMARY_VEC_SIZE  = INPUT_PRIMARY_VEC_NUM_BYTES / sizeof(T);
             constexpr uint32_t INPUT_PRIMARY_INT4_SIZE = INPUT_PRIMARY_VEC_NUM_BYTES / sizeof(int4);
 
-            const int offset_num_groups = expert_idx * num_tokens_per_expert * hidden_dim_num_groups
-                                          + token_idx * hidden_dim_num_groups + hidden_dim_group_idx;
+            // Use ``int64_t`` so the byte offset ``offset_num_groups *
+            // GROUP_SIZE`` (used at the global ``st_global`` below) cannot
+            // wrap when ``M * hidden_dim_num_groups * GROUP_SIZE >= 2^31``.
+            // Concrete failure: long-context CP4 prefill wo_b feeds
+            // ``[M=274092, K=8192]`` -> ``offset_num_groups`` in [0, 17.5M),
+            // ``offset_num_groups * GROUP_SIZE=128`` peaks at 2.245B which
+            // overflows int32 and produces a negative pointer offset. The
+            // ``st_global`` then writes off ``output_q``'s allocation and
+            // surfaces a sticky ``CUDA_ERROR_ILLEGAL_ADDRESS`` (700) at the
+            // next sync (typically inside DeepGEMM's TMA encode /
+            // ``cuLaunchKernel``), which makes it look like a DeepGEMM bug.
+            const int64_t offset_num_groups =
+                static_cast<int64_t>(expert_idx) * num_tokens_per_expert * hidden_dim_num_groups
+                + static_cast<int64_t>(token_idx) * hidden_dim_num_groups + hidden_dim_group_idx;
 
             int4 input_primary_int4[INPUT_PRIMARY_INT4_SIZE];
             T*   input_primary_vec = reinterpret_cast<T*>(input_primary_int4);
